Renames the version1 folder to v1
