+++ api/test/test_get_orders.py
@@ -1,6 +1,6 @@
 import unittest
 import json
-from api.version1.orders import app
+from api.v1.orders import app
 
 class TestOrders(unittest.TestCase):
 	"""This class tests the fetching of orders"""
