+++ api/v1/orders/__init__.py
@@ -1,5 +1,5 @@
 from flask import Flask
 app = Flask(__name__)
 
-import api.version1.orders.views   
+import api.v1.orders.views
