+++ api/v1/orders/models.py
@@ -1,4 +1,4 @@
-from api.version1.orders import app
+from api.v1.orders import app
 from flask import request, jsonify
 
 orders_list = [
