+++ requirements.txt
@@ -25,5 +25,6 @@ pytest==3.10.0
 requests==2.20.1
 six==1.11.0
 urllib3==1.24.1
+virtualenv==16.1.0
 Werkzeug==0.14.1
 wrapt==1.10.11
